fixbug: macos

+++ src/main.py
@@ -97,7 +97,7 @@ class MulimgViewer (MulimgViewerGui):
         # open default path
         if default_path:
             try:
-                self.ImgManager.init(default_path, type=2) # one_dir_mul_img
+                self.ImgManager.init(default_path, type=2)  # one_dir_mul_img
                 self.show_img_init()
                 self.ImgManager.set_action_count(0)
                 self.show_img()
@@ -918,16 +918,16 @@ class MulimgViewer (MulimgViewerGui):
             if self.checkBox_auto_draw_color.Value:
                 # 10 colors built into the software
                 color_list = [
-                    wx.Colour(217, 26, 42, 85/100*255),
-                    wx.Colour(147, 81, 166, 65/100*255),
-                    wx.Colour(85, 166, 73, 65/100*255),
-                    wx.Colour(242, 229, 48, 95/100*255),
-                    wx.Colour(242, 116, 5, 95/100*255),
-                    wx.Colour(242, 201, 224, 95/100*255),
-                    wx.Colour(36, 132, 191, 75/100*255),
-                    wx.Colour(65, 166, 90, 65/100*255),
-                    wx.Colour(214, 242, 206, 95/100*255),
-                    wx.Colour(242, 163, 94, 95/100*255)]
+                    wx.Colour(217, 26, 42, int(85/100*255)),
+                    wx.Colour(147, 81, 166, int(65/100*255)),
+                    wx.Colour(85, 166, 73, int(65/100*255)),
+                    wx.Colour(242, 229, 48, int(95/100*255)),
+                    wx.Colour(242, 116, 5, int(95/100*255)),
+                    wx.Colour(242, 201, 224, int(95/100*255)),
+                    wx.Colour(36, 132, 191, int(75/100*255)),
+                    wx.Colour(65, 166, 90, int(65/100*255)),
+                    wx.Colour(214, 242, 206, int(95/100*255)),
+                    wx.Colour(242, 163, 94, int(95/100*255))]
                 num_box = len(self.xy_magnifier)
                 if num_box <= len(color_list):
                     self.color_list = color_list[0:num_box]
@@ -1251,7 +1251,7 @@ class MulimgViewer (MulimgViewerGui):
 
     def title_auto_fc(self, event):
         titles = [self.title_down_up, self.title_show_parent,
-                  self.title_show_name, self.title_show_suffix, self.title_show_prefix,self.title_exif]
+                  self.title_show_name, self.title_show_suffix, self.title_show_prefix, self.title_exif]
         if self.title_auto.Value:
             for title in titles:
                 title.Enabled = False
